+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Breaking change: The `--from=all` output now includes the `License-Expression` value
 * Fixed KeyError with `--partial` and `--allow-only` if a license matches multiple allowed licenses.
 * Declare support for Python 3.13 and 3.14
+* Added RST/Sphinx workflow example for `--with-license-file` option in documentation
 
 ### 5.0.0